backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ryan Anderson <ryan@anderson.vc>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.60.2"
uvicorn = "^0.11.8"
python-multipart = "^0.0.5"
requests = "^2.24.0"
aiofiles = "^0.5.0"
jinja2 = "^2.11.2"
python-dotenv = "^0.14.0"
openai = "^0.2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
requests = "^2.24.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
